♻️ extract session directory
